@@ -1,13 +1,19 @@
 {
   "name": "monster-rolodex",
+  "homepage": "https://cristianstudor.github.io/monster-rolodex",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://cristianstudor.github.io/monster-rolodex",
   "dependencies": {
-    "gh-pages": "^2.1.1",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-scripts": "3.1.0"
+    "gh-pages": "^4.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "react-scripts": "5.0.1",
+    "@svgr/webpack": "^6.3.1"
+  },
+  "overrides": {
+    "@svgr/webpack": "$@svgr/webpack"
   },
   "scripts": {
     "start": "react-scripts start",
