@@ -1,6 +1,6 @@
 import React from 'react';
-import { CardList } from './components/card-list/card-list.component';
-import { SearchBox } from './components/search-box/search-box.component';
+import { CardList } from '../components/card-list/card-list.component';
+import { SearchBox } from '../components/search-box/search-box.component';
 import './App.css';
 
 class App extends React.Component {
